Moved after_upload.unzip_shapefile() to util.unzip_shapefile() so it can be used elsewhere
